# backend/open-shop-Api-spec/open-shop-api-spec.yaml
openapi: 3.0.3
info:
  title: Open Shop API
  description: E-commerce platform API for Open Shop application
  version: 1.0.0
  contact:
    name: Open Shop API Support
    email: api-support@openshop.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.openshop.com/v1
    description: Production server
  - url: https://staging-api.openshop.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
...
security:
  - bearerAuth: [ ]
paths:
...
  /admin/inventory:
    get:
      tags:
        - admin
      summary: List all inventory items with stock details
      description: |
        Retrieve all products with detailed inventory information.
        Supports filtering by stock levels, categories, and suppliers.

        **Pagination** — cursor-based via response headers:
        - First request: omit `cursor`.
        - The response sets `X-Next-Cursor` (cursor for the next page) and `X-Has-More` (`true`/`false`).
        - Next request: pass the `X-Next-Cursor` value as `?cursor=<value>`.
        - Stop when `X-Has-More: false` — there is no `X-Next-Cursor` on the last page.
        - The cursor is currently the ID of the last inventory item in the returned list.

        Requires admin JWT Bearer token with inventory management permissions.
        Rate limited to 200 requests per minute per admin user.
      operationId: listInventory
      security:
        - adminBearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - name: limit
          in: query
          description: Maximum number of inventory items to return per request (default 20, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: lowStock
          in: query
          description: Filter products with stock below minimum threshold
          schema:
            type: boolean
            example: true
        - name: categoryId
          in: query
          description: Filter by category ID
          schema:
            $ref: '#/components/schemas/ID'
        - name: supplierId
          in: query
          description: Filter by supplier ID
          schema:
            $ref: '#/components/schemas/ID'
        - name: colorFamily
          in: query
          description: Filter by color family
          schema:
            type: string
            enum: [ black, white, gray, red, blue, green, yellow, orange, purple, pink, brown, navy, rose_gold, mint, coral, burgundy, multicolor ]
            example: black
        - name: searchTerm
          in: query
          description: Search in product name, SKU, or colorway name
          schema:
            type: string
            maxLength: 100
            example: "Bluetooth black"
        - name: sortBy
          in: query
          description: Sort results by field
          schema:
            type: string
            enum: [ name, stockQuantity, lastUpdated, sellingPrice, colorway.name ]
            default: name
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [ asc, desc ]
            default: asc
      responses:
        '200':
          description: Inventory list retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-Next-Cursor:
              $ref: '#/components/headers/NextCursorHeader'
            X-Has-More:
              $ref: '#/components/headers/HasMoreHeader'
            Link:
              $ref: '#/components/headers/PaginationLinkHeader'
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Request limit per time window
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Requests remaining in current time window
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/inventory/{id}:
    get:
      tags:
        - admin
      summary: Get detailed inventory information for a product
      description: |
        Retrieve comprehensive inventory details for a specific product including
        stock levels, movements, supplier information, and cost data.

        Requires admin JWT Bearer token with inventory management permissions.
        Rate limited to 500 requests per minute per admin user.
      operationId: getInventoryDetail
      security:
        - adminBearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: Product identifier
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Inventory details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductInventory'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - admin
      summary: Update inventory details for a product
      description: |
        Update stock levels, minimum thresholds, supplier information, and costs.
        Changes are tracked in the audit log with admin user attribution.

        Requires admin JWT Bearer token with inventory management permissions.
        Rate limited to 100 requests per minute per admin user.
      operationId: updateInventory
      security:
        - adminBearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: Product identifier
          schema:
            $ref: '#/components/schemas/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryUpdateRequest'
      responses:
        '200':
          description: Inventory updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductInventory'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/inventory/{id}/adjust:
    post:
      tags:
        - admin
      summary: Perform manual stock adjustment
      description: |
        Make manual adjustments to stock levels with reason codes and notes.
        All adjustments are logged with admin attribution and timestamps.
        Supports both positive and negative adjustments.

        Requires admin JWT Bearer token with inventory management permissions.
        Rate limited to 50 requests per minute per admin user.
      operationId: adjustStock
      security:
        - adminBearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: Product identifier
          schema:
            $ref: '#/components/schemas/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockAdjustmentRequest'
      responses:
        '200':
          description: Stock adjustment completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockAdjustmentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/inventory/{id}/history:
    get:
      tags:
        - admin
      summary: Get stock movement history for a product
      description: |
        Retrieve complete audit trail of stock movements, adjustments, and changes
        for a specific product with filtering.

        **Pagination** — cursor-based via response headers:
        - First request: omit `cursor`.
        - The response sets `X-Next-Cursor` (cursor for the next page) and `X-Has-More` (`true`/`false`).
        - Next request: pass the `X-Next-Cursor` value as `?cursor=<value>`.
        - Stop when `X-Has-More: false` — there is no `X-Next-Cursor` on the last page.
        - The cursor is currently the ID of the last movement record in the returned list.

        Requires admin JWT Bearer token with inventory management permissions.
        Rate limited to 200 requests per minute per admin user.
      operationId: getStockHistory
      security:
        - adminBearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: Product identifier
          schema:
            $ref: '#/components/schemas/ID'
        - $ref: '#/components/parameters/CursorParam'
        - name: limit
          in: query
          description: Maximum number of movement records to return per request (default 50, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: fromDate
          in: query
          description: Filter movements from this date
          schema:
            type: string
            format: date-time
            example: "2024-01-01T00:00:00Z"
        - name: toDate
          in: query
          description: Filter movements to this date
          schema:
            type: string
            format: date-time
            example: "2024-12-31T23:59:59Z"
        - name: movementType
          in: query
          description: Filter by movement type
          schema:
            type: string
            enum: [ adjustment, sale, restock, return, damage, audit ]
      responses:
        '200':
          description: Stock history retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-Next-Cursor:
              $ref: '#/components/headers/NextCursorHeader'
            X-Has-More:
              $ref: '#/components/headers/HasMoreHeader'
            Link:
              $ref: '#/components/headers/PaginationLinkHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockHistoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/inventory/low-stock:
    get:
      tags:
        - admin
      summary: Get products with low stock levels
      description: |
        Retrieve products that have stock levels below their minimum threshold.
        Useful for generating reorder reports and inventory alerts.

        **Pagination** — cursor-based via response headers:
        - First request: omit `cursor`.
        - The response sets `X-Next-Cursor` (cursor for the next page) and `X-Has-More` (`true`/`false`).
        - Next request: pass the `X-Next-Cursor` value as `?cursor=<value>`.
        - Stop when `X-Has-More: false` — there is no `X-Next-Cursor` on the last page.
        - The cursor is currently the ID of the last product in the returned list.

        Requires admin JWT Bearer token with inventory management permissions.
        Rate limited to 100 requests per minute per admin user.
      operationId: getLowStockProducts
      security:
        - adminBearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - name: limit
          in: query
          description: Maximum number of products to return per request (default 20, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: severity
          in: query
          description: Filter by stock level severity
          schema:
            type: string
            enum: [ critical, low, warning ]
            example: critical
        - name: colorFamily
          in: query
          description: Filter by color family
          schema:
            type: string
            enum: [ black, white, gray, red, blue, green, yellow, orange, purple, pink, brown, navy, rose_gold, mint, coral, burgundy, multicolor ]
            example: black
        - name: productId
          in: query
          description: Filter by specific product ID
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Low stock products retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-Next-Cursor:
              $ref: '#/components/headers/NextCursorHeader'
            X-Has-More:
              $ref: '#/components/headers/HasMoreHeader'
            Link:
              $ref: '#/components/headers/PaginationLinkHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LowStockResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/inventory/bulk-update:
    post:
      tags:
        - admin
      summary: Bulk update inventory for multiple products
      description: |
        Update inventory details for multiple products in a single operation.
        Supports CSV upload or JSON payload with validation and error reporting.
        All changes are logged with admin attribution.

        Requires admin JWT Bearer token with inventory management permissions.
        Rate limited to 10 requests per minute per admin user.
      operationId: bulkUpdateInventory
      security:
        - adminBearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkInventoryUpdateRequest'
      responses:
        '200':
          description: Bulk update completed with results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkInventoryUpdateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/inventory/alerts:
    get:
      tags:
        - admin
      summary: Get inventory alerts and notifications
      description: |
        Retrieve active inventory alerts including low stock, overstock,
        and unusual movement patterns requiring admin attention.

        Requires admin JWT Bearer token with inventory management permissions.
        Rate limited to 100 requests per minute per admin user.
      operationId: getInventoryAlerts
      security:
        - adminBearerAuth: [ ]
      parameters:
        - name: alertType
          in: query
          description: Filter by alert type
          schema:
            type: string
            enum: [ low_stock, overstock, unusual_movement, expired, damaged ]
        - name: severity
          in: query
          description: Filter by alert severity
          schema:
            type: string
            enum: [ critical, high, medium, low ]
        - name: acknowledged
          in: query
          description: Filter by acknowledgment status
          schema:
            type: boolean
            example: false
      responses:
        '200':
          description: Inventory alerts retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryAlertsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # Admin - Order Management Endpoints
  /admin/orders:
    get:
      tags:
        - admin
      summary: List all customer orders with advanced filtering
      description: |
        Retrieve all customer orders with comprehensive filtering options.
        Includes order details, customer information, and payment status.

        **Pagination** — cursor-based via response headers:
        - First request: omit `cursor`.
        - The response sets `X-Next-Cursor` (cursor for the next page) and `X-Has-More` (`true`/`false`).
        - Next request: pass the `X-Next-Cursor` value as `?cursor=<value>`.
        - Stop when `X-Has-More: false` — there is no `X-Next-Cursor` on the last page.
        - The cursor is currently the ID of the last order in the returned list.

        Requires admin JWT Bearer token with order management permissions.
        Rate limited to 200 requests per minute per admin user.
      operationId: listAllOrders
      security:
        - adminBearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - name: limit
          in: query
          description: Maximum number of orders to return per request (default 20, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: status
          in: query
          description: Filter by order status
          schema:
            type: string
            enum: [ pending, confirmed, processing, shipped, delivered, cancelled ]
        - name: customerId
          in: query
          description: Filter by specific customer ID
          schema:
            $ref: '#/components/schemas/ID'
        - name: fromDate
          in: query
          description: Filter orders from this date
          schema:
            type: string
            example: "2024-01-01"
        - name: toDate
          in: query
          description: Filter orders to this date
          schema:
            type: string
            example: "2024-01-31"
        - name: minAmount
          in: query
          description: Filter orders with minimum total amount
          schema:
            type: number
            format: float
            minimum: 0
            example: 50.00
        - name: maxAmount
          in: query
          description: Filter orders with maximum total amount
          schema:
            type: number
            format: float
            minimum: 0
            example: 500.00
        - name: searchTerm
          in: query
          description: Search in order number, customer name, or email
          schema:
            type: string
            maxLength: 100
            example: "john.doe@example.com"
        - name: sortBy
          in: query
          description: Sort orders by field
          schema:
            type: string
            enum: [ createdAt, updatedAt, totalAmount, orderNumber ]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [ asc, desc ]
            default: desc
      responses:
        '200':
          description: Orders retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-Next-Cursor:
              $ref: '#/components/headers/NextCursorHeader'
            X-Has-More:
              $ref: '#/components/headers/HasMoreHeader'
            Link:
              $ref: '#/components/headers/PaginationLinkHeader'
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Request limit per time window
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Requests remaining in current time window
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/orders/{id}:
    get:
      tags:
        - admin
      summary: Get detailed order information
      description: |
        Retrieve comprehensive details for a specific order including customer information,
        items, payment details, shipping information, and order history.

        Requires admin JWT Bearer token with order management permissions.
        Rate limited to 500 requests per minute per admin user.
      operationId: getAdminOrderDetail
      security:
        - adminBearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: Order identifier
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          description: Order details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - admin
      summary: Update order status and details
      description: |
        Update order status, shipping information, or add admin notes.
        Status changes trigger customer notifications and inventory updates.

        Requires admin JWT Bearer token with order management permissions.
        Rate limited to 100 requests per minute per admin user.
      operationId: updateOrderStatus
      security:
        - adminBearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: Order identifier
          schema:
            $ref: '#/components/schemas/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminOrderUpdateRequest'
      responses:
        '200':
          description: Order updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Admin - Sales Analytics Endpoints
  /admin/sales/summary:
    get:
      tags:
        - admin
      summary: Get sales summary and key metrics
      description: |
        Retrieve high-level sales metrics including total revenue, order count,
        average order value, and growth comparisons for dashboard display.

        Requires admin JWT Bearer token with analytics permissions.
        Rate limited to 100 requests per minute per admin user.
      operationId: getSalesSummary
      security:
        - adminBearerAuth: [ ]
      parameters:
        - name: period
          in: query
          description: Time period for summary
          schema:
            type: string
            enum: [ today, yesterday, last_7_days, last_30_days, this_month, last_month, this_year ]
            default: last_30_days
        - name: compareWith
          in: query
          description: Period to compare with for growth metrics
          schema:
            type: string
            enum: [ previous_period, same_period_last_year ]
            default: previous_period
      responses:
        '200':
          description: Sales summary retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/sales/daily:
    get:
      tags:
        - admin
      summary: Get daily sales data for charts and trends
      description: |
        Retrieve daily sales data for generating charts, trends, and analytics.
        Includes revenue, order count, and average order value by day.

        Requires admin JWT Bearer token with analytics permissions.
        Rate limited to 100 requests per minute per admin user.
      operationId: getDailySales
      security:
        - adminBearerAuth: [ ]
      parameters:
        - name: fromDate
          in: query
          required: true
          description: Start date for data range
          schema:
            type: string
            example: "2024-01-01"
        - name: toDate
          in: query
          required: true
          description: End date for data range
          schema:
            type: string
            example: "2024-01-31"
        - name: groupBy
          in: query
          description: Group data by time period
          schema:
            type: string
            enum: [ day, week, month ]
            default: day
      responses:
        '200':
          description: Daily sales data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailySalesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/sales/top-products:
    get:
      tags:
        - admin
      summary: Get best-selling products analytics
      description: |
        Retrieve top-selling products by quantity sold, revenue generated,
        or profit margins within a specified time period.

        Requires admin JWT Bearer token with analytics permissions.
        Rate limited to 100 requests per minute per admin user.
      operationId: getTopProducts
      security:
        - adminBearerAuth: [ ]
      parameters:
        - name: period
          in: query
          description: Time period for analysis
          schema:
            type: string
            enum: [ last_7_days, last_30_days, this_month, last_month, this_year ]
            default: last_30_days
        - name: sortBy
          in: query
          description: Sort products by metric
          schema:
            type: string
            enum: [ quantity_sold, revenue, profit, order_count ]
            default: revenue
        - name: limit
          in: query
          description: Number of top products to return
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Top products data retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopProductsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/customers:
    get:
      tags:
        - admin
      summary: List all customers with filtering and search
      description: |
        Retrieve all customer accounts with comprehensive filtering options and search capabilities.

        **Pagination** — cursor-based via response headers:
        - First request: omit `cursor`.
        - The response sets `X-Next-Cursor` (cursor for the next page) and `X-Has-More` (`true`/`false`).
        - Next request: pass the `X-Next-Cursor` value as `?cursor=<value>`.
        - Stop when `X-Has-More: false` — there is no `X-Next-Cursor` on the last page.
        - The cursor is currently the ID of the last customer in the returned list.

        Requires admin JWT Bearer token with customer management permissions.
        Rate limited to 200 requests per minute per admin user.
      operationId: listAllCustomers
      security:
        - adminBearerAuth: [ ]
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - name: limit
          in: query
          description: Maximum number of customers to return per request (default 20, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: searchTerm
          in: query
          description: Search in customer name, email, or phone
          schema:
            type: string
            maxLength: 100
        - name: status
          in: query
          description: Filter by customer status
          schema:
            type: string
            enum: [ active, inactive, suspended ]
        - name: emailVerified
          in: query
          description: Filter by email verification status
          schema:
            type: boolean
        - name: registeredFrom
          in: query
          description: Filter customers registered from this date
          schema:
            type: string
        - name: registeredTo
          in: query
          description: Filter customers registered to this date
          schema:
            type: string
        - name: sortBy
          in: query
          description: Sort customers by field
          schema:
            type: string
            enum: [ createdAt, lastLogin, totalOrders, totalSpent ]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [ asc, desc ]
            default: desc
      responses:
        '200':
          description: Customers retrieved successfully
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestIdHeader'
            X-Next-Cursor:
              $ref: '#/components/headers/NextCursorHeader'
            X-Has-More:
              $ref: '#/components/headers/HasMoreHeader'
            Link:
              $ref: '#/components/headers/PaginationLinkHeader'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCustomerListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication

    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT Bearer token with elevated permissions

    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: |
        HTTP-only SameSite=Strict refresh token cookie. Set automatically on login, register,
        OAuth callback, and token refresh. Used by web clients to refresh access tokens without
        exposing the refresh token to JavaScript.
  parameters:
    CursorParam:
      name: cursor
      in: query
      required: false
      description: |
        Opaque cursor token for pagination. Omit on the first request.
        Use the value from the `X-Next-Cursor` response header to fetch the next page.
        Currently the cursor is the ID of the last item in the previous page.
      schema:
        type: string
        example: "65f1a2b3c4d5e6f7g8h9i0j1"
  headers:
    RequestIdHeader:
      description: Unique request identifier for tracing and debugging
      schema:
        type: string
        pattern: '^req-[a-zA-Z0-9]+$'
        example: "req-abc123def456ghi789"
    NextCursorHeader:
      description: |
        Cursor token for the next page. Pass this as the `cursor` query parameter on the
        next request. Absent when there are no more items (`X-Has-More: false`).
        Currently the value is the ID of the last item in the returned list.
      schema:
        type: string
        example: "65f1a2b3c4d5e6f7g8h9i0j9"
    HasMoreHeader:
      description: Whether more items exist beyond the current page.
      schema:
        type: string
        enum: [ "true", "false" ]
        example: "true"
    PaginationLinkHeader:
      description: |
        RFC 5988 Link header pointing to the next page URL. Present only when
        `X-Has-More: true`. Provided for clients that prefer the standard Link header.
      schema:
        type: string
        example: '<https://api.openshop.com/v1/products?cursor=65f1a2b3c4d5e6f7g8h9i0j9&limit=20>; rel="next"'
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    ID:
      type: string
      description: Identifier — can be numeric (Oracle sequence) or UUID v4
      example: "12345"
...