# backend/open-shop-webserver/modules/admin/handler.go
package admin

import (
	"net/http"

	genadmin "open-shop-webserver/gen/admin"
)

type AdminHandler struct{}

var _ genadmin.ServerInterface = (*AdminHandler)(nil)

func NewHandler() *AdminHandler {
	return &AdminHandler{}
}

func (h *AdminHandler) ListAllCustomers(w http.ResponseWriter, r *http.Request, params genadmin.ListAllCustomersParams) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) ListInventory(w http.ResponseWriter, r *http.Request, params genadmin.ListInventoryParams) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) GetInventoryAlerts(w http.ResponseWriter, r *http.Request, params genadmin.GetInventoryAlertsParams) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) BulkUpdateInventory(w http.ResponseWriter, r *http.Request) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) GetLowStockProducts(w http.ResponseWriter, r *http.Request, params genadmin.GetLowStockProductsParams) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) GetInventoryDetail(w http.ResponseWriter, r *http.Request, id genadmin.ID) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) UpdateInventory(w http.ResponseWriter, r *http.Request, id genadmin.ID) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) AdjustStock(w http.ResponseWriter, r *http.Request, id genadmin.ID) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) GetStockHistory(w http.ResponseWriter, r *http.Request, id genadmin.ID, params genadmin.GetStockHistoryParams) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) ListAllOrders(w http.ResponseWriter, r *http.Request, params genadmin.ListAllOrdersParams) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) GetAdminOrderDetail(w http.ResponseWriter, r *http.Request, id genadmin.ID) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) UpdateOrderStatus(w http.ResponseWriter, r *http.Request, id genadmin.ID) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) GetDailySales(w http.ResponseWriter, r *http.Request, params genadmin.GetDailySalesParams) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) GetSalesSummary(w http.ResponseWriter, r *http.Request, params genadmin.GetSalesSummaryParams) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}

func (h *AdminHandler) GetTopProducts(w http.ResponseWriter, r *http.Request, params genadmin.GetTopProductsParams) {
	http.Error(w, "not implemented", http.StatusNotImplemented)
}